% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_benchmark.R
\name{add_benchmark}
\alias{add_benchmark}
\title{Add benchmark to plot}
\usage{
add_benchmark(
  plot,
  y_int,
  plot_linetype = "dashed",
  plot_linecolor = OBI.color::prim_pink(),
  plot_linewidth = 0.8,
  add_label = F,
  label_text = NA,
  label_size = 5,
  label_xval = NA,
  label_yval = NA
)
}
\arguments{
\item{plot}{ggplot object}

\item{y_int}{Yintercept for geom_hline}

\item{plot_linetype}{Type of line for benchmark; default is "dashed"}

\item{plot_linecolor}{Color of line for benchmark (and benchmark label, if included); default is OBI.color::prim_pink()}

\item{plot_linewidth}{Width of line for benchmark; default is "0.8}

\item{add_label}{Whether or not benchmark should be labelled; default is FALSE}

\item{label_text}{Text for benchmark label}

\item{label_size}{Size of label text; default is 5}

\item{label_xval}{x position for benchmark label}

\item{label_yval}{y position for benchmark label}
}
\description{
this function adds a benchmark line (and optional label) to a ggplot object
}
