% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_CI_values.R
\name{add_CI_values}
\alias{add_CI_values}
\title{confidence interval for a rate}
\usage{
add_CI_values(data, var, n_cases, mean_or_proportion = "proportion")
}
\arguments{
\item{var}{variable you want to sort}

\item{n_cases}{number of cases used to calculate CI}

\item{mean_or_proportion}{Confidence interval for a mean or proportion? Set to proportion by default}
}
\description{
This function create upper and lower CI for observed mean or proportion.
Two new variables will be added: LC and UC
the formula is based on 95\% CI for a PROPORTION: rate+/- 1.96 x sqrt((!!var x (1-!!var))/!!n_cases) OR 95\% CI for a MEAN: mean +/- (1.96 * (sd(!!var, na.rm = T) / sqrt(!!n_cases)))
}
\examples{
library(dplyr)
tb = tibble::tibble(cs_rate = c(0.1, 0.3),
                    n_pt_cases = c(140, 234))
tb \%>\% 
 add_CI_values(var = cs_rate,
               n_cases = n_pt_cases)
               
OME_dt |> 
  add_CI_values(var = opioid_OME_total,
                n_cases = n_pt,
                mean_or_proportion = "mean")
               
}
