% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_muti_selection_var.R
\name{add_indicator_flgs_based_on_multiselect_var}
\alias{add_indicator_flgs_based_on_multiselect_var}
\title{Add individual procedure indicator flag vars}
\usage{
add_indicator_flgs_based_on_multiselect_var(
  data,
  map = NULL,
  multiselect_var = NULL,
  var_suffix = "_smm_cdc_flg"
)
}
\arguments{
\item{data}{OBI data}

\item{map}{indication map; e.g. see data(smm_cmp_indications_map)}

\item{multiselect_var}{variable name; e.g. severe_procedure_e, variables ends with _e}

\item{var_suffix}{indication variable name suffix, e.g. "_flg"}
}
\description{
Add individual procedure indicator flag vars
}
\details{
This function takes a muti-select variable for example severe_procedure_e and creates flags for each indication
The value to label mapping is defined in *_indications_map.csv
}
\examples{
/dontrun{
obi_dt = readRDS("/Volumes/nur-kanelow/OBI_abstracted_data/Current_Data/data/output/merged_add_data_sources.Rds")
data(smm_cmp_indications_map)
add_indicator_flgs_based_on_multiselect_var(obi_dt,
                                            map = smm_cmp_indications_map,
                                            severe_procedure_e,
                                            "_smm_cdc_flg")
}

}
