% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p4p_measures_complaince.R
\name{average_days_to_submit}
\alias{average_days_to_submit}
\title{Function to calculate the average days to submit for each site}
\usage{
average_days_to_submit(obi_dt)
}
\arguments{
\item{obi_dt}{A data frame containing the necessary columns (patientid, site_name, deliv_to_submit_max_days_int)}
}
\value{
A data frame with the average days to submit for each site
}
\description{
This function takes OBI data as input and calculates the average days to submit for each site.
It selects the patient ID, site name, and maximum days from delivery to submission from the input data frame.
Then, it calculates the average days to submit for each site and returns the result as a data frame.
}
\examples{
data <- data.frame(patientid = c(1, 2, 3),
                   site_name = c("Site A", "Site B", "Site A"),
                   deliv_to_submit_max_days_int = c(5, 7, 3))
average_days_to_submit(data)

}
\seealso{
Other {2024 P4P measures}: 
\code{\link{prop_scheduled_non_opioid_meds}()},
\code{\link{pv_email_submission_rate}()},
\code{\link{race_ethnicity_measure}()}
}
\concept{{2024 P4P measures}}
