% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incident_cases.R
\name{calculate_collab_incident_cases}
\alias{calculate_collab_incident_cases}
\title{Calculate collaborative-wide incident cases}
\usage{
calculate_collab_incident_cases(df, current_year)
}
\arguments{
\item{df}{Dataframe to be used for calculation}

\item{current_year}{What year of data is the "current" year?}
}
\value{
A single number representing the total incident cases across the collaborative. This number is NEGATIVE if there were fewer cesareans this year than last and POSITIVE if there were more cesareans this year than last.
}
\description{
Calculate collaborative-wide incident cases
}
\seealso{
Other incident_cases: 
\code{\link{create_incident_case_site_df}()}
}
\concept{incident_cases}
