% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_dys_comp.R
\name{calculate_dys_comp}
\alias{calculate_dys_comp}
\title{Calculate dystocia compliance}
\usage{
calculate_dys_comp(df, include_types = F)
}
\arguments{
\item{df}{A data frame, preferably one that has already been run through obstint::create_obi_cohort()}

\item{include_types}{If TRUE, calculates compliance for individual types of dystocia}
}
\description{
To calculate dystocia compliance, overall and by type
}
