% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dystocia_comp.R
\name{calculate_dys_comp_rate}
\alias{calculate_dys_comp_rate}
\title{Dystocia compliance functions
To calculate dystocia compliance, overall and by type}
\usage{
calculate_dys_comp_rate(df, include_types = F)
}
\arguments{
\item{df}{A data frame, preferably one that has already been run through obstint::create_obi_cohort()}

\item{include_types}{If TRUE, calculates compliance for individual types of dystocia}
}
\description{
Dystocia compliance functions
To calculate dystocia compliance, overall and by type
}
