% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_IA_comp.R
\name{calculate_IA_comp}
\alias{calculate_IA_comp}
\title{Calculate IA compliance}
\usage{
calculate_IA_comp(df)
}
\arguments{
\item{df}{Data frame -- usually this is the data frame read into R using data.table::fread}
}
\description{
Calculate IA compliance
}
