% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_max_submitdate.R
\name{calculate_max_submitdate}
\alias{calculate_max_submitdate}
\title{Calculate maximum submitdate}
\usage{
calculate_max_submitdate()
}
\arguments{
\item{df}{Data frame -- usually this is obi_cohort created with obstinit::read_current_data() |> obstinit::create_obi_cohort()}
}
\description{
this function creates a variable called "submitdate_max" that is the maximum of the other submitdate fields in the dataset. This is primarily used for tracking completeness and timeliness of abstracted data per the 2023 and 2024 OBI scorecards.
}
