% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_TB_comp.R
\name{calculate_TB_comp}
\alias{calculate_TB_comp}
\title{Calculate TeamBirth compliance}
\usage{
calculate_TB_comp(df)
}
\arguments{
\item{df}{Data frame -- note that in order to have the correct variables for numerator and denominator calculation, data frame MUST first be run through obstinit::create_teambirth_cohort}
}
\description{
Calculate TeamBirth compliance
}
