% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cat_II_cohort.R
\name{create_cat_II_cohort}
\alias{create_cat_II_cohort}
\title{Create cat II cohort}
\usage{
create_cat_II_cohort(df, limit_to_catII_sites = F)
}
\arguments{
\item{df}{Data frame -- usually this is the data frame read into R using data.table::fread}

\item{limit_to_catII_sites}{Whether or not to limit the cohort to 2023 Cat II sites - set to false by default}
}
\description{
Create cat II cohort
}
