% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_caterpillar_df.R
\name{create_caterpillar_df}
\alias{create_caterpillar_df}
\title{Create a data frame for caterpillar plots}
\usage{
create_caterpillar_df(df, num_var, denom_var, group_var = site_name)
}
\arguments{
\item{df}{Data frame -- usually this is the data frame read into R using data.table::fread}

\item{limit_to_locked}{Logical -- if TRUE, only return cases that are locked}
}
\description{
this function creates a data frame to be used for caterpillar plots where each row is a site name. The df includes a numerator, denominator, and rate for the variable of interest, as well as a 95\% CI (bounded by 0 and 1) for the rate.
}
