% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dystocia_cohort.R
\name{create_dystocia_cohort}
\alias{create_dystocia_cohort}
\title{Create dystocia cohort}
\usage{
create_dystocia_cohort(df)
}
\arguments{
\item{df}{A data frame, preferably one that has already been run through obstint::create_obi_cohort()}
}
\description{
To create dystocia compliance cohort
}
