% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_IA_cohort.R
\name{calculate_IA_comp}
\alias{calculate_IA_comp}
\title{Create IA cohort}
\usage{
calculate_IA_comp(df, limit_to_IA_sites = F)
}
\arguments{
\item{df}{Data frame -- usually this is the data frame read into R using data.table::fread}

\item{limit_to_IA_sites}{Whether cohort should be limited to 2023 IA sites; default is false}
}
\description{
Create IA cohort
}
