% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incident_cases.R
\name{create_incident_case_site_df}
\alias{create_incident_case_site_df}
\title{Create incident case site df}
\usage{
create_incident_case_site_df(df, current_year)
}
\arguments{
\item{df}{Dataframe to be used for calculation}

\item{current_year}{What year of data is the "current" year?}
}
\value{
A dataframe with one row per site, containing each sites' incident cases comparing the current year to last year. Grouped by MDHHS ID.
}
\description{
Create incident case site df
}
\seealso{
Other incident_cases: 
\code{\link{calculate_collab_incident_cases}()}
}
\concept{incident_cases}
