% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_obi_cohort.R
\name{create_obi_cohort}
\alias{create_obi_cohort}
\title{Create OBI cohort of locked, complete cases for most analyses}
\usage{
create_obi_cohort(df)
}
\arguments{
\item{df}{Data frame -- usually this is the data frame read into R using data.table::fread}
}
\description{
Create OBI cohort of locked, complete cases for most analyses
}
