% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project_folder.R
\name{create_project_folder}
\alias{create_project_folder}
\title{Automatically create folders for a project}
\usage{
create_project_folder()
}
\description{
Automatically create folders for a project
}
\details{
Automatically generate a project
folder structure and .ignore in the working directory of your R session.
}
