% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient_voice.R
\name{create_pv_response_dt}
\alias{create_pv_response_dt}
\title{create patient voice response data}
\usage{
create_pv_response_dt(obi_dt, pro_survey_expired = FALSE)
}
\arguments{
\item{obi_dt}{OBI nightly export data; use current data, becasue expiration date is based on current date}

\item{pro_survey_expired}{logical, if TRUE, only include patients who are after 12 weeks postpartum}
}
\value{
a data frame with variables: patientid, site_name, infant_dob_dt, complete_pro_survey_flg, survey_informed_consent_b_0
}
\description{
This function create OBI subset cohort of pts who received pro survey emails.
There are two additional vars added to OBI dataset
\itemize{
\item pro_response: pts who completed pro survey pro_response == 1
\item survey_informed_consent_b_0: pts choose to opt out ==0; Self opt out pts also consider complete the survey.
}
}
\details{
this function read in pro survey data from
\itemize{
\item survey response: \verb{P:/OBI_abstracted_data/Current_Data/data/input/pro_6_week_postpartum.csv}
\item contact log pts who received emails: \verb{P:/OBI_abstracted_data/Current_Data/data/input/contactlog.csv}
}
}
\examples{
\dontrun{
obi_dt <- read_current_data() |> 
  create_obi_cohort(limit_to_locked = FALSE) 
create_pv_response_dt(obi_dt) 
}
 
}
