% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_teambirth_cohort.R
\name{create_teambirth_cohort}
\alias{create_teambirth_cohort}
\title{Create TeamBirth cohort}
\usage{
create_teambirth_cohort(df, limit_to_teambirth_sites = T)
}
\arguments{
\item{df}{Data frame -- usually this is obi_cohort created using obstinit::create_obi_cohort}

\item{limit_to_teambirth_sites}{Whether cohort should be limited to 2023 teambirth sites; default is TRUE}
}
\description{
Create TeamBirth cohort
}
