% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_website_upload_csv.R
\name{create_website_upload_csv}
\alias{create_website_upload_csv}
\title{Create website upload CSV}
\usage{
create_website_upload_csv(
  report_name,
  members_only = 1,
  site_year = 2023,
  exclude_site = c(""),
  output_path
)
}
\arguments{
\item{report_name}{A character string that is the tile of the report excluding the site names. The report filename must be formatted as "\link{filename} - \link{sitename}".}

\item{members_only}{A logical value that indicates whether the report is for members only. Default is 1.}

\item{site_year}{An integer that indicates the year used to determine the sites that have reports. Default is 2023.}

\item{exclude_site}{A character vector that contains any sites to exclude. Default is an empty vector.}

\item{output_path}{Path to output the CSV file to.}
}
\description{
To create the CSV needed to upload reports to the OBI website.
}
