% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_paths.R
\name{dropbox_root_path}
\alias{dropbox_root_path}
\title{Create Dropbox root path}
\usage{
dropbox_root_path(pc_unique_name = "ourdea")
}
\arguments{
\item{pc_unique_name}{your unique name. This will be used to find your account on PC.
No need to input this if using Mac}
}
\description{
Based in PC or Mac. If on PC, specify the unique name of the PC user.
}
