% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_paths.R
\name{gdrive_root_path}
\alias{gdrive_root_path}
\title{OBI Google Drive root path}
\usage{
gdrive_root_path(uniqname)
}
\arguments{
\item{uniqname}{uniqname to use for mac users}
}
\description{
Find Google Drive root path based on whether machine is Mac or PC and username (if Mac)
}
