% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_gt.R
\name{gt_obi}
\alias{gt_obi}
\title{Format GT table}
\usage{
gt_obi(
  gt_obj,
  heading.title.font.size = 14,
  heading.subtitle.font.size = 12,
  column_labels.font.size = 12,
  row_group.font.size = 12,
  stub.font.size = 12,
  table.font.size = 12,
  footnotes.font.size = 10
)
}
\arguments{
\item{gt_obj}{A gt object to be formatted.}

\item{heading.title.font.size}{Size for table title. Default is 14.}

\item{heading.subtitle.font.size}{Size for table subtitle. Default is 12.}

\item{column_labels.font.size}{Size for column labels. Default is 12.}

\item{row_group.font.size}{Size for row group labels. Default is 12.}

\item{stub.font.size}{Size for stub labels. Default is 12.}

\item{table.font.size}{Size for table content. Default is 12.}

\item{footnotes.font.size}{Size for footnotes. Default is 10.}
}
\description{
Format GT table
}
