% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iol_ebp_missingness.R
\name{iol_ebp_missingness}
\alias{iol_ebp_missingness}
\title{Create website upload CSV}
\usage{
iol_ebp_missingness(main_data, cervical_data, induction_data, group = F, ...)
}
\arguments{
\item{main_data}{The input dataset containing the main wide dataset or analytictable.}

\item{cervical_data}{The input dataset containing the Cervical and Amniotomy cangrow.}

\item{induction_data}{The input dataset containing the Induction cangrow.}

\item{group}{A logical value that indicates whether the data should be grouped. Default is FALSE.}

\item{...}{Grouping arguments}
}
\value{
If group = F, a dataset contain patient-level missingness counts is returned. If group = T, a dataset with the group-level missingness counts and rate is returend.
}
\description{
To create the CSV needed to upload reports to the OBI website. The CSV files should be in the following \href{https://docs.google.com/spreadsheets/d/19Ua83FMIsT9RjrugrvPD4NJEZcfI8e_cBFfrudDSDeM/edit#gid=141729511}{format}
(see the \href{https://docs.google.com/document/d/1g0BcXkA-rcm0AAtkTjK9uE7Bh2o5NzB_fBr_UDr_niw/edit}{SOP} for more details). If you are uploading a report multiple times in a month and would like to overwrite past versions,
you do not need to re-run this function; you can reuse the previous CSV as long as the report name remains unchanged. If you do NOT want to overwrite previous versions uploaded in the same month, you will need to ensure
the report does not have the same name as the previous version.
}
\examples{
/dontrun{
library(tidyverse)

test <- iol_ebp_missingness(obi_26_dt, cerv_cangrow, iol_cangrow)

grouped_test <- iol_ebp_missingness(obi_26_dt, cerv_cangrow, iol_cangrow, group = T, site_name)
}

}
