% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nna_progress.R
\name{NNA_progress}
\alias{NNA_progress}
\title{NNA Progress}
\usage{
NNA_progress(
  obi_dt,
  start_date = "2024-10-01",
  end_date = "2025-10-01",
  OE_output = "OE_ratios/2025/current_OE_ratio_dt.xlsx",
  ...,
  imputed_24 = FALSE
)
}
\arguments{
\item{obi_dt}{OBI nightly export data.}

\item{start_date}{The start date of the time period for which you are interested in seeing NNA progress. Must be formatted as "YYYY-MM-DD". Default is "2024-10-01".}

\item{end_date}{The end date of the time period for which you are interested in seeing NNA progress. Must be formatted as "YYYY-MM-DD". Default is "2025-10-01".}

\item{OE_output}{The .csv file path to the O:E output. Default is the path for the 2025 O:E output: "OE_ratios/2025/current_OE_ratio_dt.xlsx". The turbo root path is automatically included.}

\item{...}{Grouping arguments. By default includes external_mdhhs_site_id and site_name.}

\item{imputed_24}{Logical. If TRUE, the baseline cesarean rate will use imputed rates where available. Default is FALSE.}
}
\value{
A dataframe with the progress towards the NNA goals during the selected timeframe.
}
\description{
To calculate progress towards the NNA goal.
}
\examples{
/dontrun{
NNA_output <- NNA_progress(
obi_dt, 
start_date = "2024-10-01", 
end_date = "2024-11-10", 
OE_output, 
g1_status_25, delivery_volume_2023_lbl
)
}

}
