% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_caterpillar_chart.R
\name{plot_caterpillar_chart}
\alias{plot_caterpillar_chart}
\title{Plot caterpillar chart}
\usage{
plot_caterpillar_chart(
  df,
  y_var = rate,
  x_var = site_name,
  point_shape = "diamond",
  point_size = 2,
  dt_color = OBI.color::prim_dark_blue()
)
}
\arguments{
\item{df}{Data frame created by create_caterpillar_df()}

\item{y_var}{Variable to be placed on the y axis - default is rate, which is the rate variable created by create_caterpillar_df()}

\item{x_var}{Variable to be placed on the x axis - default is site_name}

\item{point_shape}{Shape of the points - default is "diamond"}

\item{point_size}{Size of the points - default is 2}

\item{dt_color}{Color for the data points and errorbars - default is OBI.color::prim_dark_blue()}
}
\description{
this function plots a caterpillar chart using the dataframe created by create_caterpillar_df()
}
