% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ctrl_hchart.R
\name{plot_ctrl_hchart}
\alias{plot_ctrl_hchart}
\title{Plot interactive ctrl chart using highcharter package}
\usage{
plot_ctrl_hchart(df, title, ymin = NA, ymax = NA)
}
\arguments{
\item{df}{A data frame - should be in the structure of obstinit::structure_data}

\item{title}{Title for plot}
}
\description{
Plot interactive ctrl chart using highcharter package
}
