% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_obi_report.R
\name{print_obi_report}
\alias{print_obi_report}
\title{Create website upload CSV}
\usage{
print_obi_report(
  input_path,
  output_pdf_path,
  top_margin = 0.4,
  bottom_margin = 0.4,
  page_number = FALSE,
  page_range = ""
)
}
\arguments{
\item{input_path}{A character string that is the file path of the html report.}

\item{output_pdf_path}{A character string that is the file path of the final PDF report.}

\item{top_margin}{A numeric that indicates the top margin of the PDF. Default is 0.4.}

\item{bottom_margin}{A numeric that indicates the bottom margin of the PDF. Default is 0.4.}

\item{page_number}{A logical that indicates if the page number should be included in the PDF. Default is FALSE.}

\item{page_range}{A character string that indicates the page range of the PDF. Default is "", which prings all pages.}
}
\value{
Creates a PDF containing the BCBSM logo and language in a footnote. The file is a PDF copy of the HTML verson of the report..
}
\description{
To create the CSV needed to upload reports to the OBI website. The CSV files should be in the following \href{https://docs.google.com/spreadsheets/d/19Ua83FMIsT9RjrugrvPD4NJEZcfI8e_cBFfrudDSDeM/edit#gid=141729511}{format}
(see the \href{https://docs.google.com/document/d/1g0BcXkA-rcm0AAtkTjK9uE7Bh2o5NzB_fBr_UDr_niw/edit}{SOP} for more details)
}
\examples{
/dontrun{
print_obi_report("report.html", 
"PRO Patient Voices Report.pdf", 
top_margin = 0.4, 
bottom_margin = 0.4,
page_number = TRUE)
}

}
