% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p4p_measures_complaince.R
\name{race_ethnicity_measure}
\alias{race_ethnicity_measure}
\title{Calculate race and ethnicity missing measure}
\usage{
race_ethnicity_measure(
  obi_dt,
  ...,
  limit_to_2024 = T,
  by_site = lifecycle::deprecated()
)
}
\arguments{
\item{obi_dt}{Dataframe to be piped into the function
#' @param ... Grouping arguments.}

\item{limit_to_2024}{Logical value indicating whether to filter the data to cases with infant dob year ≥ 2024; default is set to T}

\item{by_site}{Should the output dataframe be parsed by site? Default is T. DEPRECATED JANUARY 2025.}
}
\value{
A summarized data table with the calculated measures.
}
\description{
Calculate race and ethnicity missing measure
}
\details{
It then calculates the number of patients, the number of missing
race and ethnicity values, and the percentage of missing values. If \code{by_site} is TRUE,
the measures are calculated separately for each site.
}
\examples{
/dontrun{
# Calculate race and ethnicity measures for the entire dataset
race_ethnicity_measure(obi_dt)

# Calculate race and ethnicity measures by site
race_ethnicity_measure(obi_dt, by_site = TRUE)
}
}
\seealso{
Other {2024 P4P measures}: 
\code{\link{average_days_to_submit}()},
\code{\link{prop_births_mtg_COMFORT_compliance}()},
\code{\link{prop_scheduled_non_opioid_meds}()},
\code{\link{pv_email_submission_rate}()}
}
\concept{{2024 P4P measures}}
