% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_crosswalk.R
\name{read_crosswalk}
\alias{read_crosswalk}
\title{Read in OBI crosswalk}
\usage{
read_crosswalk()
}
\description{
this function reads the current OBI crosswalk file. This file contains information on hospital variables like teaching status, as well as group 1 status and color.
}
