% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_current_data.R
\name{read_current_data}
\alias{read_current_data}
\title{Read current OBI data}
\usage{
read_current_data(sas_processed_dt = TRUE)
}
\arguments{
\item{sas_processed_dt}{if read SAS processed data, default to TRUE}
}
\description{
read OBI analytic data from Turbo
}
