% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ecqm_data.R
\name{read_ecqm_data}
\alias{read_ecqm_data}
\title{Read current eCQM data}
\usage{
read_ecqm_data(limit_to_denominator = TRUE)
}
\arguments{
\item{limit_to_denominator}{if TRUE then dataset is limited to only eCQM denominator cases. Default is TRUE.}
}
\description{
read eCQM data from Turbo
}
