% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_email.R
\name{send_email}
\alias{send_email}
\title{send emails from R}
\usage{
send_email(
  email_message = "<p> This is an automatic message. </p> <p> OBI analytics team </p> ",
  email_to = "obianalysts@umich.edu",
  email_from = NULL,
  email_subject = "test",
  cc = NULL,
  attachment = NULL,
  reply_to = NULL,
  save_or_send = "save"
)
}
\arguments{
\item{email_message}{your email body; use html code for breaks and hyperlinks}

\item{email_to}{an email address or a list of addresses in format: c(";")}

\item{email_from}{email send from address, for example \code{OBICustomerSupport@med.umich.edu}}

\item{email_subject}{your email subject}

\item{cc}{an email address or a list of addresses in format: c(";")}

\item{attachment}{a path name for a file to be attached to the email}

\item{reply_to}{an email address to which all replies will be sent}

\item{save_or_send}{default to "save", can be "save" or "send"}
}
\value{
a message in the console
}
\description{
Send or save draft emails from R
}
\examples{
\dontrun{
send_email(email_message = "<p> This is a test.")
}
}
