% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{site_names}
\alias{site_names}
\title{Site name list for website upload}
\format{
\subsection{\code{site_names}}{

A data frame with 1 column:
\describe{
\item{site_name}{character string of site name}
}
}
}
\usage{
site_names
}
\description{
Used in create_website_upload_csv() as a list of site names. Update 11/07/24 to change Ascensions to Henry Fords. Update 2/14/25 to correct Marshfield Dickinson.
}
\keyword{datasets}
