% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{site_names_mdhhs}
\alias{site_names_mdhhs}
\title{Site MDHHS ID list}
\format{
\subsection{\code{site_names}}{

A data frame with 3 columns:
\describe{
\item{site_name}{character string of site name}
\item{site_id_a_mx}{double AMx ID}
\item{mdhhs_id}{character string MDHHS ID}
}
}
}
\usage{
site_names_mdhhs
}
\description{
Used in create_website_upload_csv() to map site names to MDHHS IDs. Update 11/07/24 to change Ascensions to Henry Fords. Update 2/14/25 to correct Marshfield Dickinson. Update 6/4/25 to add McLaren Petoskey.
}
\keyword{datasets}
