% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sites_w_all_years_dt.R
\name{sites_w_all_years_dt}
\alias{sites_w_all_years_dt}
\title{Limit dataset to hospitals that have data for all years 2020 - max year of your choice}
\usage{
sites_w_all_years_dt(df, max_year)
}
\arguments{
\item{df}{Data frame -- usually this is OBI's current, processed data read into R using data.table::fread. Preferably this has already been run through obstinit::create_obi_cohort()}
}
\description{
Limit dataset to hospitals that have data for all years 2020 - max year of your choice
}
