% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smm_cmp_indications_map}
\alias{smm_cmp_indications_map}
\title{SMM complication indications map}
\format{
\subsection{\code{smm_cmp_indications_map}}{

A data frame with 2 columns:
\describe{
\item{value}{numeric value}
\item{indication}{value label text}
}
}
}
\usage{
smm_cmp_indications_map
}
\description{
Used in nightly export data processing to create separate indicator variables for each SMM complication indication.
}
\keyword{datasets}
