% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_muti_selection_var.R
\name{sort_muti_selection}
\alias{sort_muti_selection}
\title{sort multiple selected varible as list}
\usage{
sort_muti_selection(var)
}
\arguments{
\item{var}{variable you want to sort}
}
\description{
multiple selected varibales shouldn't have different values based on order
for example, {1|2} euqal to {2|1}. This function short the numbers
}
\examples{
val = "8,1,0,3"
sort_muti_selection(val)
}
