% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_muti_selection_var.R
\name{sort_muti_selection_var}
\alias{sort_muti_selection_var}
\title{sort multiple selected variable in dataset}
\usage{
sort_muti_selection_var(var)
}
\arguments{
\item{var}{}
}
\description{
multiple selected varibales shouldn't have different values based on order
for example, {1|2} euqal to {2|1}. This function short the numbers
}
\examples{
df <- tibble(col = c("{1|2}", "{2|1}", "{3|2|1}", "{2|3|1}", "{2|1|3}"))
df \%>\% mutate(new = sort_muti_selection_var(col))
}
