% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_mdhhs_id_5.R
\name{standardize_mdhhs_id_5}
\alias{standardize_mdhhs_id_5}
\title{Normalize MDHHS IDs to 5-digit format}
\usage{
standardize_mdhhs_id_5(id)
}
\arguments{
\item{id}{A character vector or numeric vector of MDHHS IDs.}
}
\value{
A character vector of normalized 5-digit MDHHS IDs.
}
\description{
This function normalizes Michigan Department of Health and Human Services (MDHHS) IDs to a standard 5-digit format.
\itemize{
\item If the ID has 4 digits, a leading zero is added.
\item If the ID has 6 digits, the first digit is removed. (SS format)
\item IDs with 5 digits are left unchanged.
}
}
\examples{
tibble(id = c("1234", "01234", "101234", "56789", "123456")) |>
  mutate(mdhhs_id = standardize_mdhhs_id_5(id))

}
