% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_data.R
\name{structure_data}
\alias{structure_data}
\title{Structure OBI data for plotting control charts}
\usage{
structure_data(
  df,
  date_var,
  date_gran = c(year_mon, year_qtr),
  num_var,
  den_var
)
}
\arguments{
\item{df}{A data frame}

\item{date_var}{The date variable to be used for grouping; usually infant_dob_dt}

\item{date_gran}{The granularity of dates we want to use for our control chart; monthly or quarterly}

\item{num_var}{The variable to be summarized as the numerator of the rate we're interested in calculating - should be binary 0 1}

\item{den_var}{The variable to be summarized as the denominator of the rate we're interested in calculating - should be binary 0 1}
}
\description{
Structure OBI data for plotting control charts
}
