% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_obi.R
\name{theme_obi}
\alias{theme_obi}
\title{Add OBI ggplot theme}
\usage{
theme_obi(
  x_axis_title_size = 12,
  x_axis_text_size = 10,
  y_axis_title_size = 12,
  y_axis_text_size = 10,
  title_size = 16,
  title_hjust = 0.5,
  subtitle_size = 12,
  subtitle_hjust = 0.5,
  caption_size = 12,
  caption_hjust = 0,
  legend_title_size = 12,
  legend_text_size = 11
)
}
\arguments{
\item{x_axis_title_size}{Size for x axis title. Default is 12.}

\item{x_axis_text_size}{Size for x axis text. Default is 10.}

\item{y_axis_title_size}{Size for y axis title. Default is 12.}

\item{y_axis_text_size}{Size for y axis text. Default is 10.}

\item{title_size}{Size for title text. Default is 16.}

\item{title_hjust}{Position for title text. Default is 0.5 for centered, change to 0 for left-justified.}

\item{subtitle_size}{Size of subtitle text. Default is 12.}

\item{subtitle_hjust}{Position for subtitle text. Default is 0.5 for centered, change to 0 for left-justified.}

\item{caption_size}{Size for caption text. Default is 12.}

\item{caption_hjust}{Position for caption. Default is 0 for left-justified, change to 0.5 for centered.}

\item{legend_title_size}{Size of legend title text. Default is 12.}

\item{legend_text_size}{Size of legend text. Default is 11.}
}
\description{
Add OBI ggplot theme
}
