% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cat_II_comp.R
\name{calculate_cat_II_comp}
\alias{calculate_cat_II_comp}
\title{Calculate cat II compliance}
\usage{
calculate_cat_II_comp(df)
}
\arguments{
\item{df}{Data frame}
}
\description{
Calculate cat II compliance
}
