% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_IA_comp.R
\name{calculate_IA_comp}
\alias{calculate_IA_comp}
\title{Calculate IA compliance}
\usage{
calculate_IA_comp(df, limit_to_IA_sites = F)
}
\arguments{
\item{df}{Data frame -- ideally, run through obstinit::create_IA_cohort first}
}
\description{
Calculate IA compliance
}
