% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_TB_comp.R
\name{calculate_TB_comp}
\alias{calculate_TB_comp}
\title{Calculate TeamBirth compliance}
\usage{
calculate_TB_comp(df)
}
\arguments{
\item{df}{Data frame -- usually this is the data frame read into R using data.table::fread}
}
\description{
Calculate TeamBirth compliance
}
