% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_obi_cohort.R
\name{create_obi_cohort}
\alias{create_obi_cohort}
\title{Create OBI cohort of locked, complete cases for most analyses}
\usage{
create_obi_cohort(df, limit_to_locked = T)
}
\arguments{
\item{df}{Data frame -- usually this is the data frame read into R using data.table::fread}

\item{limit_to_locked}{Logical -- if TRUE, only return cases that are locked}
}
\description{
this function does
\enumerate{
\item include case ≥2020-01-01 based on infant date of birth
\item include case if case is complete using flg_complete
\item create case_locked varaible using 90 days lock
\item filter to complete and locked cases(limit_to_locked = TRUE)
\item create mdhhs_id character variable with leading 0's
}
}
