% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient_voice.R
\name{create_pv_response_dt}
\alias{create_pv_response_dt}
\title{create patient voice response data}
\usage{
create_pv_response_dt(obi_dt, pro_survey_expired = FALSE)
}
\arguments{
\item{obi_dt}{OBI nightly export data}

\item{pro_survey_expired}{logical, if TRUE, only include patients who are after 16 weeks postpartum}
}
\value{
a data frame with variables: patientid, site_name, infant_dob_dt, pro_response
}
\description{
this function creates a data frame excluding opt out patients
\itemize{
\item opt out from CDA
\item opt out from patients
\item limited to infant dob > 2023-5-10 (when PRO survey started)
}
}
\details{
this function read in pro survey data from \verb{P:/OBI_abstracted_data/Current_Data/data/input/pro_6_week_postpartum.csv}
}
\examples{
obi_dt <- read_current_data() |>
  create_obi_cohort(limit_to_locked = FALSE)
create_pv_response_dt(obi_dt)
}
