% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient_voice.R
\name{create_pv_response_dt}
\alias{create_pv_response_dt}
\title{create patient voice response data}
\usage{
create_pv_response_dt(obi_dt)
}
\arguments{
\item{obi_dt}{OBI nightly export data}
}
\value{
a data frame with variables: patientid, site_name, infant_dob_dt, pro_response
}
\description{
this function creates a data frame excluding opt out patients
\itemize{
\item opt out from CDA
\item opt out from patients
}
}
\examples{
obi_dt <- read_current_data() |> 
  create_obi_cohort()
create_pv_response_dt(obi_dt)
}
