% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_website_upload_csv.R
\name{create_website_upload_csv}
\alias{create_website_upload_csv}
\title{Create website upload CSV}
\usage{
create_website_upload_csv(
  report_name,
  tags,
  members_only = 1,
  site_list,
  output_path
)
}
\arguments{
\item{report_name}{A character string that is the tile of the report excluding the site names. The report filename must be formatted as "filename - sitename".}

\item{tags}{A character vector that contains the tags for the report. The available tags must be one of the following: Data Request, NTSV Performance Report, NTSV Summary Report, Annual P4P Scorecards, PRIME Reports, Other, or Site-Specific Patient Voices Resources.}

\item{members_only}{A logical value that indicates whether the report is for members only. Default is 1.}

\item{site_list}{A character vector that contains site names for hospitals with reports.}

\item{output_path}{Path to output the CSV file to (must end in /).}
}
\value{
A saved CSV file that can be uploaded to the OBI website.
}
\description{
To create the CSV needed to upload reports to the OBI website. The CSV files should be in the following \href{https://docs.google.com/spreadsheets/d/19Ua83FMIsT9RjrugrvPD4NJEZcfI8e_cBFfrudDSDeM/edit#gid=141729511}{format}
(see the \href{https://docs.google.com/document/d/1g0BcXkA-rcm0AAtkTjK9uE7Bh2o5NzB_fBr_UDr_niw/edit}{SOP} for more details). If you are uploading a report multiple times in a month and would like to overwrite past versions,
you do not need to re-run this function; you can reuse the previous CSV as long as the report name remains unchanged. If you do NOT want to overwrite previous versions uploaded in the same month, you will need to ensure
the report does not have the same name as the previous version.
}
\examples{
/dontrun{
library(tidyverse)

site_list <- c("Michigan Medicine Ann Arbor", "Hurley Medical Center Flint")

create_website_upload_csv(report_name = "test report - Michigan Medicine",
                         tags = "NTSV Performance Report",
                          site_list = site_list,
                          output_path = "test_report.csv")
}

}
