% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_gt.R
\name{gt_obi}
\alias{gt_obi}
\title{Format GT table}
\usage{
gt_obi(
  gt_obj,
  full_width = FALSE,
  heading.title.font.size = 14,
  heading.subtitle.font.size = 12,
  column_labels.font.size = 12,
  row_group.font.size = 12,
  stub.font.size = 12,
  table.font.size = 12,
  footnotes.font.size = 10
)
}
\arguments{
\item{gt_obj}{A gt object to be formatted.}

\item{full_width}{Logical indicating whether to set the table width to 100\%. Default is FALSE}

\item{heading.title.font.size}{Size for table title. Default is 14.}

\item{heading.subtitle.font.size}{Size for table subtitle. Default is 12.}

\item{column_labels.font.size}{Size for column labels. Default is 12.}

\item{row_group.font.size}{Size for row group labels. Default is 12.}

\item{stub.font.size}{Size for stub labels. Default is 12.}

\item{table.font.size}{Size for table content. Default is 12.}

\item{footnotes.font.size}{Size for footnotes. Default is 10.}
}
\description{
Format GT table
}
\examples{
library(gt)
df <- data.frame(
  Name = c("Alice", "Bob", "Charlie"),
  Age = c(25, 30, 35),
  Score = c(90, 85, 95)
)

df |>
  gt() |>
  gt_obi(full_width = TRUE) |> 
  # title
  tab_header(
    title = "Student Information",
   subtitle = "A brief overview of students' details"
 ) 
}
