% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nna_progress.R
\name{NNA_progress}
\alias{NNA_progress}
\title{NNA Progress}
\usage{
NNA_progress(obi_dt, start_date, end_date, OE_output, ...)
}
\arguments{
\item{obi_dt}{OBI nightly export data.}

\item{start_date}{The start date of the time period for which you are interested in seeing NNA progress. Must be formatted as "YYYY-MM-DD".}

\item{end_date}{The end date of the time period for which you are interested in seeing NNA progress. Must be formatted as "YYYY-MM-DD".}

\item{OE_output}{A dataframe that contains the data read in from the O:E ratio .csv file.}

\item{...}{Grouping arguments. By default includes external_mdhhs_site_id and site_name.}
}
\value{
A dataframe with the progress towards the NNA goals during the selected timeframe.
}
\description{
To calculate progress towards the NNA goal.
}
\examples{
/dontrun{
NNA_progress(
obi_dt, 
start_date = "2024-10-01", 
end_date = "2024-11-10", 
OE_output, 
g1_status_25, delivery_volume_2023_lbl
)
}

}
