% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ctrl_chart.R
\name{plot_ctrl_chart}
\alias{plot_ctrl_chart}
\title{Plot control chart}
\usage{
plot_ctrl_chart(df)
}
\arguments{
\item{df}{A data frame in long format, created by the structure_data function}
}
\description{
Plot control chart
}
