% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p4p_measures_complaince.R
\name{prop_scheduled_non_opioid_meds}
\alias{prop_scheduled_non_opioid_meds}
\title{Function to calculate compliance with scheduled acetaminophen and oral NSAID}
\usage{
prop_scheduled_non_opioid_meds(
  obi_dt,
  ...,
  limit_to_2024 = T,
  by_site = lifecycle::deprecated()
)
}
\arguments{
\item{obi_dt}{A data frame containing the necessary columns (patientid, site_name, deliv_to_submit_max_days_int)}

\item{...}{Grouping arguments. By default includes opioid_group; other variables can be added as needed.}

\item{limit_to_2024}{Logical value indicating whether to filter the data to cases with infant dob year ≥ 2024; default is set to T}

\item{by_site}{Should the dataframe be parse by site? Default to YES since for ease of P4P scoring. DEPRECTAED OCTOBER 2024.}
}
\value{
A data frame with the proportion of eligible births with scheduled acetaminophen and oral NSAID per site
}
\description{
This function takes OBI data as input and calculates what proportion of eligible births got scheduled acetaminophen and oral NSAID
}
\examples{
data <- data.frame(patientid = c(1, 2, 3),
                   site_name = c("Site A", "Site B", "Site A"),
                   deliv_to_submit_max_days_int = c(5, 7, 3))
prop_scheduled_non_opioid_meds(data)

}
\seealso{
Other {2024 P4P measures}: 
\code{\link{average_days_to_submit}()},
\code{\link{prop_births_mtg_COMFORT_compliance}()},
\code{\link{pv_email_submission_rate}()},
\code{\link{race_ethnicity_measure}()}
}
\concept{{2024 P4P measures}}
