% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p4p_measures_complaince.R
\name{pv_email_submission_rate}
\alias{pv_email_submission_rate}
\title{Calculate the email submission rate for PRO}
\usage{
pv_email_submission_rate(obi_dt, ..., by_site = lifecycle::deprecated())
}
\arguments{
\item{obi_dt}{The input dataset containing patient information. If you only need a certain infant dob
range, you can filter the dataset before passing it to this function.}

\item{...}{Grouping arguments.}

\item{by_site}{A logical value indicating whether to calculate the email submission rate by site. DEPRECATED JANUARY 2025.}
}
\value{
A data frame with the following columns:
\itemize{
\item n_pt: The total number of patients in the dataset.
\item n_submitted_email: The number of patients who submitted their email.
\item n_submitted_pct: The percentage of patients who submitted their email.
}
}
\description{
This function calculates the email submission rate  by
checking if variable \code{email_txt} has value.
It returns the number of patients, the number of patients who submitted their email,
and the percentage of patients who submitted their email.
compliance definition is at https://docs.google.com/spreadsheets/d/123r8ShMQuxl7GQQsX2Buuh9ynOyMRHOaSeKoFHiwx58/edit#gid=793851872
}
\examples{
/dontrun{
obi_dt <- read_current_data(sas_processed_dt = FALSE)
obi_dt_subset <- obi_dt |>
   filter(infant_dob_ymd >= "2023-01-01")
pv_email_submission_rate(obi_dt_subset)
}

}
\seealso{
Other {2024 P4P measures}: 
\code{\link{average_days_to_submit}()},
\code{\link{prop_births_mtg_COMFORT_compliance}()},
\code{\link{prop_scheduled_non_opioid_meds}()},
\code{\link{race_ethnicity_measure}()}
}
\concept{{2024 P4P measures}}
