% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p4p_measures_complaince.R
\name{race_ethnicity_measure}
\alias{race_ethnicity_measure}
\title{Calculate race and ethnicity missing measure}
\usage{
race_ethnicity_measure(obi_dt, by_site = TRUE)
}
\arguments{
\item{obi_dt}{The input dataset containing patient information. If you only need a certain infant dob
range, you can filter the dataset before passing it to this function.}

\item{by_site}{A logical value indicating whether to calculate the email submission rate by site.}
}
\value{
A summarized data table with the calculated measures.
}
\description{
Calculate race and ethnicity missing measure
}
\details{
It then calculates the number of patients, the number of missing
race and ethnicity values, and the percentage of missing values. If \code{by_site} is TRUE,
the measures are calculated separately for each site.
}
\examples{
/dontrun{
# Calculate race and ethnicity measures for the entire dataset
race_ethnicity_measure(obi_dt)

# Calculate race and ethnicity measures by site
race_ethnicity_measure(obi_dt, by_site = TRUE)
}

}
\seealso{
Other {2024 P4P measures}: 
\code{\link{average_days_to_submit}()},
\code{\link{pv_email_submission_rate}()}
}
\concept{{2024 P4P measures}}
