% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_current_data.R
\name{read_current_data}
\alias{read_current_data}
\title{Read current OBI data}
\usage{
read_current_data(sas_processed_dt = lifecycle::deprecated())
}
\arguments{
\item{sas_processed_dt}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
SAS processed data is not supported (SAS processing stopped in Aug 2024).}
}
\value{
A data frame containing the current OBI data.
}
\description{
Reads OBI analytic data from Turbo.
}
