% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cat_II_comp.R, R/read_current_data.R
\name{calculate_cat_II_comp}
\alias{calculate_cat_II_comp}
\alias{read_current_data}
\title{Calculate cat II compliance}
\usage{
calculate_cat_II_comp(df)

read_current_data()
}
\arguments{
\item{df}{Data frame -- usually this is the data frame read into R using data.table::fread}
}
\description{
Calculate cat II compliance

Read current data
}
