% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ecqm_data.R
\name{read_ecqm_data}
\alias{read_ecqm_data}
\title{Read current eCQM data}
\usage{
read_ecqm_data(risk_adjust = FALSE)
}
\arguments{
\item{risk_adjust}{if reading only risk adjustment data; default to FALSE}
}
\description{
read eCQM data from Turbo
}
