% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_email.R
\name{send_email}
\alias{send_email}
\title{send emails from R}
\usage{
send_email(
  email_message = "<p> This is an automatic message. </p> <p> OBI analytics team </p> ",
  email_to = "obianalysts@umich.edu",
  email_subject = "test",
  cc = NULL,
  save_or_send = "save"
)
}
\arguments{
\item{email_message}{your email body; use html code for breaks and hyperlinks}

\item{email_to}{a email address or a list of email addresses separated by ";"}

\item{email_subject}{text}

\item{cc}{a email address or a list of email addresses separated by ";"}

\item{save_or_send}{default to "save", can be "save" or "send"}
}
\value{
a message in the console
}
\description{
send or save draft emails from R
}
\examples{
send_email(email_message = "<p> This is an automatic message. </p> <p> OBI analytics team </p>")
}
